refactor: Entity PK 자료형 Long으로 변경
현재 db에 저장되는 PK 자료형인 UNSIGNED INT의 범위는 Java의 Integer.MAX_VALUE를
초과하기 때문에 자료형을 Long으로 변경했습니다.

Issue:

+++ backend/main/src/main/java/com/cobby/main/costume/api/controller/CostumeController.java
@@ -49,7 +49,7 @@ public class CostumeController {
 	public ResponseEntity<? extends BaseResponseBody> getCostume(
 		@PathVariable
 		@Positive(message = "올바르지 않은 ID 양식입니다.") // 최대 상한 조건(Max)도 추가를 고려해야 함
-		Integer costumeId) {
+		Long costumeId) {
 
 		var costume = costumeService.selectCostume(costumeId);
 
@@ -82,7 +82,7 @@ public class CostumeController {
 	public ResponseEntity<? extends BaseResponseBody> deleteCostume(
 		@PathVariable
 		@Positive(message = "올바르지 않은 ID 양식입니다.") // 최대 상한 조건(Max)도 추가를 고려해야 함
-		Integer costumeId) {
+		Long costumeId) {
 
 		var deletedId = costumeService.deleteCostume(costumeId);
