+++ backend/main/src/main/java/com/cobby/main/quest/api/dto/response/QuestGetResponse.java
@@ -17,7 +17,7 @@ import java.util.List;
 @AllArgsConstructor
 public class QuestGetResponse {
 
-    private Integer questId;
+    private Long questId;
 
     private String questName;
