+++ backend/main/src/main/java/com/cobby/main/quest/db/repository/QuestRepository.java
@@ -9,6 +9,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface QuestRepository extends JpaRepository<Quest, Integer> {
+public interface QuestRepository extends JpaRepository<Quest, Long> {
 	List<Quest> findAllByQuestTypeOrderByQuestGoal(QuestCategory questCategory);
 }
